cleanups

+++ src/module.F90
@@ -415,7 +415,7 @@ end module parameters
 module MemoryAllocator
 !-------------------------------------------------------------------------------------------
 implicit none
-integer :: totalMemory=0
+integer(8) :: totalMemory=0
 
 contains 
 
@@ -575,7 +575,7 @@ subroutine DeallocatorI3D(array)
   return
 end subroutine 
 
-integer function GetTotalMemory() 
+integer(8) function GetTotalMemory() 
   GetTotalMemory = totalMemory
   return
 end function
